Theorem: gnome-score-helper was (while not perfect) better the way it was.
Reasoning:
	. gnome-score-helper is intended to allow games to change
	scores, but not users.
	. Therefore, it must be given additional permissions
	beyond what a user normally has.
	. Therefore, it must be either suid or sgid.
	. Nobody currently has the knowledge (or the time) to do a
	complete audit of libgnome and glib, so we cannot ensure
	the security of this program.
	. Maintainance of score files cannot be done over a
	network transport (such as might be used by gnome_config
	in the future) in a secure manner.
	. Therefore, we need a plain-jane s[ug]id score helper.

Nathan Bryant pointed out that making it suid would stop people from
ptracing it and would also help at file creation time, and assuming we can
get rid of glib dependencies, that seems to me to be a sane proposition.
This version I'm checking in (still an older version) still uses glib I
believe, but I'll try to eliminate that dependency in my copious spare
time (or feel free to do it yourself... :-)

The issue here isn't the score helper (which is probably not very
consequential), but whether we are going to put the extra effort into
making Gnome as security-conscious as possible. Let's not repeat the
mistakes of other environments such as CDE & KDE, both of which have had
serious holes at one time or another...

Leave it be unless you have a better solution, please!
-- Elliot