Added Michael Zucchi's ultra fast terminal emulation widget.
Added Michael Zucchi's ultra fast terminal emulation widget.

See sample zterm.c for an example.  We will be making a full
GNOME terminal in gnome-core.

There is a problem: it uses forkpty() which is available on
GNU libc systems and recent BSD systems, but not anywere else, we
need a compatibility routine for this.

Miguel.